integrate greedyselection into program; remove DumbProtocolSelection
